+++ Modules/CalcOffence-3_0.lua
@@ -2102,7 +2102,7 @@ function calcs.offence(env, actor, activeSkill)
 				t_insert(breakdown.ImpaleModifier, s_format("%d ^8(number of stacks, can be overridden in the Configuration tab)", impaleStacks))
 				t_insert(breakdown.ImpaleModifier, s_format("x %.3f ^8(stored damage)", impaleStoredDamage))
 				t_insert(breakdown.ImpaleModifier, s_format("x %.2f ^8(impale chance)", impaleChance))
-				t_insert(breakdown.ImpaleModifier, s_format("= %.3f", impaleDMGModifier))
+				t_insert(breakdown.ImpaleModifier, s_format("= %.3f ^8(more damage)", impaleDMGModifier))
 
 			end
 		end
@@ -2219,8 +2219,20 @@ function calcs.offence(env, actor, activeSkill)
 		output.CombinedDPS = output.CombinedDPS + output.DecayDPS
 	end
 	if skillFlags.impale then
-		output.ImpaleHit = (output.MainHand.PhysicalHitAverage + (output.OffHand.PhysicalHitAverage or output.MainHand.PhysicalHitAverage)) / 2 * (1-output.CritChance/100) + (output.MainHand.PhysicalCritAverage + (output.OffHand.PhysicalCritAverage or output.MainHand.PhysicalCritAverage)) / 2 * (output.CritChance/100)
+		output.ImpaleHit = ((output.MainHand.PhysicalHitAverage or output.OffHand.PhysicalHitAverage) + (output.OffHand.PhysicalHitAverage or output.MainHand.PhysicalHitAverage)) / 2 * (1-output.CritChance/100) + ((output.MainHand.PhysicalCritAverage or output.OffHand.PhysicalCritAverage) + (output.OffHand.PhysicalCritAverage or output.MainHand.PhysicalCritAverage)) / 2 * (output.CritChance/100)
 		output.ImpaleDPS = output.ImpaleHit * ((output.ImpaleModifier or 1) - 1) * output.Speed
-		output.TotalDPS = output.TotalDPS + output.ImpaleDPS
+		output.WithImpaleDPS = output.TotalDPS + output.ImpaleDPS
+		
+		if breakdown then
+			breakdown.ImpaleDPS = {}
+			t_insert(breakdown.ImpaleDPS, s_format("%.2f ^8(average physical hit)", output.ImpaleHit))
+			t_insert(breakdown.ImpaleDPS, s_format("x %.2f ^8(attack rate)", output.Speed))
+			t_insert(breakdown.ImpaleDPS, s_format("x %.2f ^8(impale modifier)", ((output.ImpaleModifier or 1) - 1)))
+			t_insert(breakdown.ImpaleDPS, s_format("= %.1f", output.ImpaleDPS))
+		end
 	end
+		
+	
+
+	
 end